+++ packages/hdwallet-provider/src/constructor/getPrivateKeys.ts
@@ -1,5 +1,5 @@
-import { PrivateKey } from "./constructorTypes/types";
-import { SigningAuthority } from "./constructorTypes/Constructor";
+import { PrivateKey } from "./types";
+import { SigningAuthority } from "./Constructor";
 
 // extract the private keys if that's the style used, or return undefined
 export const getPrivateKeys = (
