+++ packages/hdwallet-provider/src/constructor/types.ts
